+++ zmeventnotification/zmeventnotification/zmes_hook_helpers/utils.py
@@ -34,8 +34,8 @@ def rescale_polygons(xfactor, yfactor):
             newy = int(y * yfactor)
             newp.append((newx, newy))
         newps.append({'name': p['name'], 'value': newp})
-    g.logger.debug('resized polygons x={}/y={}: {}'.format(
-        xfactor, yfactor, newps),level=2)
+    g.logger.Debug(2,'resized polygons x={}/y={}: {}'.format(
+        xfactor, yfactor, newps))
     g.polygons = newps
 
 
@@ -55,11 +55,22 @@ def str_split(my_str):
     return [x.strip() for x in my_str.split(',')]
 
 
+
+# credit: https://stackoverflow.com/a/5320179
+def findWholeWord(w):
+    return re.compile(r'\b({0})\b'.format(w), flags=re.IGNORECASE).search
+
+
 # Imports zone definitions from ZM
 def import_zm_zones(mid, reason):
-    match_reason = True if g.config['only_triggered_zm_zones']=='yes' else False
+
+    match_reason = False
+    if reason:
+        match_reason = True if g.config['only_triggered_zm_zones']=='yes' else False
+    g.logger.Debug(2,'import_zm_zones: match_reason={} and reason={}'.format(match_reason, reason))
+
     url = g.config['portal'] + '/api/zones/forMonitor/' + mid + '.json'
-    g.logger.debug('Getting ZM zones using {}?username=xxx&password=yyy&user=xxx&pass=yyy'.format(url),level=2)
+    g.logger.Debug(2,'Getting ZM zones using {}?username=xxx&password=yyy&user=xxx&pass=yyy'.format(url))
     url = url + '?username=' + g.config['user']
     url = url + '&password=' + urllib.parse.quote(g.config['password'], safe='')
     url = url + '&user=' + g.config['user']
@@ -71,7 +82,7 @@ def import_zm_zones(mid, reason):
         main_handler = urllib.request.HTTPHandler()
 
     if g.config['basic_user']:
-        g.logger.debug('Basic auth config found, associating handlers',level=2)
+        g.logger.Debug(2,'Basic auth config found, associating handlers')
         password_mgr = urllib.request.HTTPPasswordMgrWithDefaultRealm()
         top_level_url = g.config['portal']
         password_mgr.add_password(None, top_level_url, g.config['basic_user'],
@@ -84,54 +95,50 @@ def import_zm_zones(mid, reason):
     try:
         input_file = opener.open(url)
     except HTTPError as e:
-        g.logger.error(f'HTTP Error in import_zm_zones:{e}')
+        g.logger.Error(f'HTTP Error in import_zm_zones:{e}')
         raise
     except Exception as e:
-        g.logger.error(f'General error in import_zm_zones:{e}')
+        g.logger.Error(f'General error in import_zm_zones:{e}')
         raise
 
     c = input_file.read()
     j = json.loads(c)
 
-    # Now lets look at reason to see if we need to 
+    # Now lets look at reason to see if we need to
     # honor ZM motion zones
 
-    reason_zones = []
-    if match_reason and reason and 'Motion:' in reason:
-        rz = reason.split('Motion:')
-        if len(rz) > 1:
-            rz = rz[1]
-            reason_zones = [x.strip() for x in rz.split(',')]
-            g.logger.debug('Found motion zones provided in alarm cause: {}'.format(reason_zones))
 
+    #reason_zones = [x.strip() for x in rz.split(',')]
+    #g.logger.Debug(1,'Found motion zones provided in alarm cause: {}'.format(reason_zones))
 
     for item in j['zones']:
-        if not match_reason or 'All' in reason_zones or item['Zone']['Name'] in reason_zones:
-            g.polygons.append({
-                'name': item['Zone']['Name'],
-                'value': str2tuple(item['Zone']['Coords'])
-            })
-            g.logger.debug('importing zoneminder polygon: {} [{}]'.format(
-                item['Zone']['Name'], item['Zone']['Coords']), level=2)
-        else:
-            g.logger.debug('dropping {} as zones in alarm cause is {}'.format(item['Zone']['Name'], reason_zones))
+        if  match_reason:
+            if not findWholeWord(item['Zone']['Name'])(reason):
+                g.logger.Debug(1,'dropping {} as zones in alarm cause is {}'.format(item['Zone']['Name'], reason))
+                continue
+        g.logger.Debug(2,'importing zoneminder polygon: {} [{}]'.format(item['Zone']['Name'], item['Zone']['Coords']))
+        g.polygons.append({
+            'name': item['Zone']['Name'],
+            'value': str2tuple(item['Zone']['Coords'])
+        })
+
 
 
 # downloaded ZM image files for future analysis
 def download_files(args):
     if g.config['wait'] > 0:
-        g.logger.info('Sleeping for {} seconds before downloading'.format(
+        g.logger.Info('Sleeping for {} seconds before downloading'.format(
             g.config['wait']))
         time.sleep(g.config['wait'])
 
-    
+
     if g.config['portal'].lower().startswith('https://'):
         main_handler = urllib.request.HTTPSHandler(context=g.ctx)
     else:
         main_handler = urllib.request.HTTPHandler()
 
     if g.config['basic_user']:
-        g.logger.debug('Basic auth config found, associating handlers',level=2)
+        g.logger.Debug(2,'Basic auth config found, associating handlers')
         password_mgr = urllib.request.HTTPPasswordMgrWithDefaultRealm()
         top_level_url = g.config['portal']
         password_mgr.add_password(None, top_level_url, g.config['basic_user'],
@@ -163,11 +170,11 @@ def download_files(args):
                     g.config['password'], safe='')
             durl = durl + '&username=' + g.config['user'] + '&password=*****'
 
-        g.logger.debug('Trying to download {}'.format(durl))
+        g.logger.Debug(1,'Trying to download {}'.format(durl))
         try:
             input_file = opener.open(url)
         except HTTPError as e:
-            g.logger.error(e)
+            g.logger.Error(e)
             raise
         with open(filename1, 'wb') as output_file:
             output_file.write(input_file.read())
@@ -182,11 +189,11 @@ def download_files(args):
                 'user'] + '&password=' + urllib.parse.quote(
                     g.config['password'], safe='')
             durl = durl + '&username=' + g.config['user'] + '&password=*****'
-        g.logger.debug('Trying to download {}'.format(durl))
+        g.logger.Debug(1,'Trying to download {}'.format(durl))
         try:
             input_file = opener.open(url)
         except HTTPError as e:
-            g.logger.error(e)
+            g.logger.Error(e)
             raise
         with open(filename2, 'wb') as output_file:
             output_file.write(input_file.read())
@@ -209,7 +216,7 @@ def download_files(args):
                 'user'] + '&password=' + urllib.parse.quote(
                     g.config['password'], safe='')
             durl = durl + '&username=' + g.config['user'] + '&password=*****'
-        g.logger.debug('Trying to download {}'.format(durl))
+        g.logger.Debug(1,'Trying to download {}'.format(durl))
         input_file = opener.open(url)
         with open(filename1, 'wb') as output_file:
             output_file.write(input_file.read())
@@ -222,7 +229,7 @@ def get_pyzm_config(args):
     config_file.read(args.get('config'))
     if config_file.has_option('general', 'pyzm_overrides'):
         pyzm_overrides = config_file.get('general', 'pyzm_overrides')
-        g.config['pyzm_overrides'] =  ast.literal_eval(pyzm_overrides) if pyzm_overrides else {}     
+        g.config['pyzm_overrides'] =  ast.literal_eval(pyzm_overrides) if pyzm_overrides else {}
 
 
 def process_config(args, ctx):
@@ -244,7 +251,7 @@ def process_config(args, ctx):
         elif t == 'float':
             return float(val)
         else:
-            g.logger.error(
+            g.logger.Error(
                 'Unknown conversion type {} for config key:{}'.format(
                     e['type'], e['key']))
             return val
@@ -255,18 +262,18 @@ def process_config(args, ctx):
             val = config_file[v['section']].get(k, v['default'])
         else:
             val = v['default']
-            g.logger.debug(
+            g.logger.Debug(1,
                 'Section [{}] missing in config file, using key:{} default: {}'
                 .format(v['section'], k, val))
 
         if val and val[0] == '!':  # its a secret token, so replace
-            g.logger.debug('Secret token found in config: {}'.format(val),level=2)
+            g.logger.Debug(2,'Secret token found in config: {}'.format(val))
             if not has_secrets:
                 raise ValueError(
                     'Secret token found, but no secret file specified')
             if secrets_file.has_option('secrets', val[1:]):
                 vn = secrets_file.get('secrets', val[1:])
-                #g.logger.debug ('Replacing {} with {}'.format(val,vn))
+                #g.logger.Debug (1,'Replacing {} with {}'.format(val,vn))
                 val = vn
             else:
                 raise ValueError(
@@ -278,7 +285,7 @@ def process_config(args, ctx):
             dval = g.config[k]
         else:
             dval = '***********'
-        #g.logger.debug ('Config: setting {} to {}'.format(k,dval))
+        #g.logger.Debug (1,'Config: setting {} to {}'.format(k,dval))
 
     # main
     try:
@@ -287,7 +294,7 @@ def process_config(args, ctx):
 
         if config_file.has_option('general', 'secrets'):
             secrets_filename = config_file.get('general', 'secrets')
-            g.logger.debug('secret filename: {}'.format(secrets_filename))
+            g.logger.Debug(1,'secret filename: {}'.format(secrets_filename))
             has_secrets = True
             secrets_file = ConfigParser(interpolation=None)
             try:
@@ -296,23 +303,21 @@ def process_config(args, ctx):
             except:
                 raise
         else:
-            g.logger.debug('No secrets file configured')
+            g.logger.Debug(1,'No secrets file configured')
         # now read config values
 
         for k, v in g.config_vals.items():
-            #g.logger.debug ('processing {} {}'.format(k,v))
+            #g.logger.Debug (1,'processing {} {}'.format(k,v))
             if k == 'secrets':
                 continue
 
             _set_config_val(k, v)
-            #g.logger.debug ("done")
-
         if g.config['allow_self_signed'] == 'yes':
             ctx.check_hostname = False
             ctx.verify_mode = ssl.CERT_NONE
-            g.logger.debug('allowing self-signed certs to work...')
+            g.logger.Debug(1,'allowing self-signed certs to work...')
         else:
-            g.logger.debug('strict SSL cert checking is on...')
+            g.logger.Debug(1,'strict SSL cert checking is on...')
 
         g.polygons = []
 
@@ -327,18 +332,22 @@ def process_config(args, ctx):
                     v = item[1]
                     if k in g.config_vals:
                         # This means its a legit config key that needs to be overriden
-                        g.logger.debug(
+                        g.logger.Debug(2,
                             '[{}] overrides key:{} with value:{}'.format(
-                                sec, k, v),level=2)
+                                sec, k, v))
                         g.config[k] = _correct_type(v,
                                                     g.config_vals[k]['type'])
                     else:
                         # This means its a polygon for the monitor
                         if not g.config['only_triggered_zm_zones'] == 'yes':
-                            g.polygons.append({'name': k, 'value': str2tuple(v)})
-                            g.logger.debug('adding polygon: {} [{}]'.format(k, v),level=2)
+                            try:
+                                g.polygons.append({'name': k, 'value': str2tuple(v)})
+                                g.logger.Debug(2,'adding polygon: {} [{}]'.format(k, v))
+                            except Exception as e:
+                                g.logger.Error('{}={} is either an invalid attribute or a malformed polygon. Error was {}. Ignoring.'.format(k,v,e))
+
                         else:
-                            g.logger.debug ('ignoring polygon: {} as only_triggered_zm_zones is true'.format(k), level=2)
+                            g.logger.Debug (2,'ignoring polygon: {} as only_triggered_zm_zones is true'.format(k))
             # now import zones if needed
             # this should be done irrespective of a monitor section
             if g.config['only_triggered_zm_zones'] == 'yes':
@@ -347,13 +356,13 @@ def process_config(args, ctx):
                 import_zm_zones(args.get('monitorid'), args.get('reason'))
 
         else:
-            g.logger.info(
+            g.logger.Info(
                 'Ignoring monitor specific settings, as you did not provide a monitor id'
             )
     except Exception as e:
-        g.logger.error('Error parsing config:{}'.format(args.get('config')))
-        g.logger.error('Error was:{}'.format(e))
-        g.logger.fatal('error: Traceback:{}'.format(traceback.format_exc()))
+        g.logger.Error('Error parsing config:{}'.format(args.get('config')))
+        g.logger.Error('Error was:{}'.format(e))
+        g.logger.Fatal('error: Traceback:{}'.format(traceback.format_exc()))
         exit(0)
 
     # Now lets make sure we take care of parameter substitutions {{}}
@@ -369,19 +378,19 @@ def process_config(args, ctx):
 
                 g.config[gk] = g.config[gk].replace('{{' + sub_var + '}}',
                                                     g.config[sub_var])
-                g.logger.debug('key [{}] is \'{}\' after substitution'.format(
-                    gk, g.config[gk]),level=2)
+                g.logger.Debug(2,'key [{}] is \'{}\' after substitution'.format(
+                    gk, g.config[gk]))
 
     # Now munge config if testing args provide
     if args.get('file'):
         g.config['wait'] = 0
         g.config['write_image_to_zm'] = 'no'
         g.polygons = []
-       
-        
+
+
     if  args.get('output_path'):
-        g.logger.debug ('Output path modified to {}'.format(args.get('output_path')))
+        g.logger.Debug (1,'Output path modified to {}'.format(args.get('output_path')))
         g.config['image_path'] = args.get('output_path')
         g.config['write_debug_image'] = 'yes'
-    
-   
+
+